check for solved cube
